+++ packages/typespec-azure-resource-manager/test/rules/arm-resource-interface.test.ts
@@ -31,12 +31,10 @@ describe("typespec-azure-resource-manager: detect non-post actions", () => {
       @doc(".")
       enum Versions {
         @doc(".")
-        @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-        @armCommonTypesVersion(Azure.ResourceManager.CommonTypes.Versions.v3)
+              @armCommonTypesVersion(Azure.ResourceManager.CommonTypes.Versions.v3)
         v2021_09_21: "2022-09-21-preview",
         @doc(".")
-        @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-        @armCommonTypesVersion(Azure.ResourceManager.CommonTypes.Versions.v4)
+              @armCommonTypesVersion(Azure.ResourceManager.CommonTypes.Versions.v4)
         v2022_01_10: "2022-01-10-alpha.1"
       }