+++ packages/typespec-azure-resource-manager/test/rules/arm-resource-name-pattern.test.ts
@@ -24,8 +24,7 @@ it("Emits a warning for an ARM resource that doesn't specify `@pattern` on the n
     .expect(
       `
       @armProviderNamespace
-      @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-      namespace Microsoft.Contoso;
+          namespace Microsoft.Contoso;
       
       model Employee is ProxyResource<{}> {
         @key("employeeName")
@@ -59,8 +58,7 @@ it("Allows codefix when ARM resource name is missing pattern.", async () => {
     .expect(
       `
       @armProviderNamespace
-      @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-      namespace Microsoft.Contoso;
+          namespace Microsoft.Contoso;
       
       model Employee is ProxyResource<{}> {
         @key("employeeName")
@@ -72,8 +70,7 @@ it("Allows codefix when ARM resource name is missing pattern.", async () => {
     )
     .applyCodeFix("add-pattern-decorator").toEqual(`
       @armProviderNamespace
-      @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-      namespace Microsoft.Contoso;
+          namespace Microsoft.Contoso;
       
       model Employee is ProxyResource<{}> {
         @pattern("^[a-zA-Z0-9-]{3,24}$")
@@ -90,8 +87,7 @@ it("Does not emit a warning for an ARM resource that specifies `@pattern` on the
     .expect(
       `
     @armProviderNamespace
-    @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-    namespace Microsoft.Contoso;
+      namespace Microsoft.Contoso;
     
     model Employee is ProxyResource<{}> {
       @pattern("^[a-zA-Z0-9-]{3,24}$")
@@ -119,8 +115,7 @@ it("Does not emit a warning for an ARM resource that specifies `@pattern` on the
     .expect(
       `
     @armProviderNamespace
-    @useDependency(Azure.ResourceManager.Versions.v1_0_Preview_1)
-    namespace Microsoft.Contoso;
+      namespace Microsoft.Contoso;
 
     @pattern("^[a-zA-Z0-9][a-zA-Z0-9-]{1,58}[a-zA-Z0-9]$")
     scalar stringResourceName extends string;