+++ packages/typespec-client-generator-core/src/lib.ts
@@ -210,7 +210,7 @@ export const $lib = createTypeSpecLibrary({
     "multiple-response-types": {
       severity: "warning",
       messages: {
-        default: paramMessage`Multiple response types found in operation ${"operation"}. Only one response type is supported, so we will choose the first one ${"response"}`,
+        default: paramMessage`Multiple response types found in operation ${"operation"}. Some emitters might not support returning all of these response types`,
       },
     },
     "no-corresponding-method-param": {
@@ -261,6 +261,13 @@ export const $lib = createTypeSpecLibrary({
       messages: {
         default: paramMessage`Client name: "${"name"}" is duplicated in language scope: "${"scope"}"`,
         nonDecorator: paramMessage`Client name: "${"name"}" is defined somewhere causing naming conflicts in language scope: "${"scope"}"`,
+      },
+    },
+    "client-name-ineffective": {
+      severity: "warning",
+      messages: {
+        default: paramMessage`Application of @clientName decorator to ${"name"} is not effective`,
+        override: paramMessage`Application of @clientName decorator to ${"name"} is not effective because it is applied to the override method. Please apply it on the original method definition "${"originalMethodName"}" instead.`,
       },
     },
     "example-loading": {