+++ packages/typespec-client-generator-core/src/public-utils.ts
@@ -41,6 +41,7 @@ import {
 } from "./decorators.js";
 import {
   SdkBodyParameter,
+  SdkClient,
   SdkClientType,
   SdkCookieParameter,
   SdkHeaderParameter,
@@ -48,6 +49,7 @@ import {
   SdkHttpOperationExample,
   SdkMethodParameter,
   SdkModelPropertyType,
+  SdkOperationGroup,
   SdkPathParameter,
   SdkQueryParameter,
   SdkServiceMethod,
@@ -847,4 +849,26 @@ export function resolveOperationId(
  */
 export function isHttpMetadata(context: TCGCContext, property: SdkModelPropertyType): boolean {
   return property.__raw !== undefined && isMetadata(context.program, property.__raw);
+}
+
+export function getNamespaceFromType(
+  type: Type | SdkClient | SdkOperationGroup | undefined,
+): Namespace | undefined {
+  if (type === undefined) {
+    return undefined;
+  }
+  if (type.kind === "SdkOperationGroup" || type.kind === "SdkClient") {
+    const rawType = type.type;
+    if (rawType === undefined) {
+      return undefined;
+    }
+    if (rawType.kind === "Namespace") {
+      return rawType;
+    }
+    return rawType.namespace;
+  }
+  if ("namespace" in type) {
+    return type.namespace;
+  }
+  return undefined;
 }